# apps/backend/docs/openapi/swagger.yaml
basePath: /api/v1
definitions:
  gin.H:
    additionalProperties: {}
    type: object
...
host: localhost:8080
info:
  contact:
    email: security@arc-hawk.io
    name: ARC-Hawk Security Team
  description: Enterprise PII Discovery, Classification, and Lineage Tracking Platform.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: https://arc-hawk.io/terms
  title: ARC-HAWK-DD API
  version: 3.0.0
paths:
...
  /discovery/reports:
    get:
      parameters:
      - description: Bearer {token}
        in: header
        name: Authorization
        required: true
        type: string
      - description: Max results (default 50)
        in: query
        name: limit
        type: integer
      - description: Offset
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/gin.H'
      security:
      - BearerAuth: []
      summary: List recent report jobs
      tags:
      - discovery
  /discovery/reports/{id}:
    get:
      parameters:
      - description: Bearer {token}
        in: header
        name: Authorization
        required: true
        type: string
      - description: Report UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/gin.H'
      security:
      - BearerAuth: []
      summary: Get report job status
      tags:
      - discovery
  /discovery/reports/{id}/download:
    get:
      description: Returns application/pdf, text/html, text/csv, or application/json
        depending on format
      parameters:
      - description: Bearer {token}
        in: header
        name: Authorization
        required: true
        type: string
      - description: Report UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/pdf
      responses:
        "200":
          description: Report bytes
          schema:
            type: string
        "425":
          description: Report not ready yet
          schema:
            $ref: '#/definitions/gin.H'
      security:
      - BearerAuth: []
      summary: Download generated report file
      tags:
      - discovery
  /discovery/reports/generate:
    post:
      consumes:
      - application/json
      description: Generates html, csv, json, or pdf report asynchronously. Poll GET
        /discovery/reports/:id for status.
      parameters:
      - description: Bearer {token}
        in: header
        name: Authorization
        required: true
        type: string
      - description: '{snapshot_id, format: html|csv|json|pdf}'
        in: body
        name: body
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "202":
          description: report_id, status
          schema:
            $ref: '#/definitions/gin.H'
      security:
      - BearerAuth: []
      summary: Enqueue a discovery report
      tags:
      - discovery
...
  /fp/check:
    post:
      consumes:
      - application/json
      parameters:
      - description: Bearer {token}
        in: header
        name: Authorization
        required: true
        type: string
      - description: '{pattern_name, field_name, asset_id}'
        in: body
        name: body
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/gin.H'
      security:
      - BearerAuth: []
      summary: Check if field/pattern is a known false positive
      tags:
      - fp-learning
  /fp/confirmed:
    post:
      consumes:
      - application/json
      parameters:
      - description: Bearer {token}
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/gin.H'
      security:
      - BearerAuth: []
      summary: Confirm a finding as true PII
      tags:
      - fp-learning
  /fp/false-positives:
    post:
      consumes:
      - application/json
      parameters:
      - description: Bearer {token}
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/gin.H'
      security:
      - BearerAuth: []
      summary: Mark a finding as false positive
      tags:
      - fp-learning
  /fp/learnings:
    get:
      parameters:
      - description: Bearer {token}
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/gin.H'
      security:
      - BearerAuth: []
      summary: List active FP learning entries
      tags:
      - fp-learning
  /fp/learnings/{id}:
    delete:
      parameters:
      - description: Bearer {token}
        in: header
        name: Authorization
        required: true
        type: string
      - description: Learning UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/gin.H'
      security:
      - BearerAuth: []
      summary: Soft-delete an FP learning entry
      tags:
      - fp-learning
    get:
      parameters:
      - description: Bearer {token}
        in: header
        name: Authorization
        required: true
        type: string
      - description: Learning UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/gin.H'
      security:
      - BearerAuth: []
      summary: Get a specific FP learning entry
      tags:
      - fp-learning
  /fp/stats:
    get:
      parameters:
      - description: Bearer {token}
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/gin.H'
      security:
      - BearerAuth: []
      summary: Get FP learning precision/recall stats
      tags:
      - fp-learning
...
securityDefinitions:
  APIKeyAuth:
    description: Tenant API key for service-to-service authentication
    in: header
    name: X-API-Key
    type: apiKey
  BearerAuth:
    description: 'JWT Bearer token. Format: "Bearer <token>"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"